# backend/assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# Nexus Tasks REST API\n\nNexus Tasks is a collaborative project and task management platform.\nThis document describes the full REST API surface of the backend service.\n\n## Authentication\n\nMost endpoints require an authenticated session established via the `/api/v2/auth/login`\nor `/api/v2/auth/register` endpoints. Authentication state is maintained via an\nHTTP-only session cookie (`nexus_user`). Alternatively, machine clients may\nauthenticate using an API Key passed in the `X-API-Key` request header.\n\n## Base URL\n\nAll versioned resources are prefixed with `/api/v2`.",
    "title": "provides the Swagger/OpenAPI specification metadata for the Nexus Tasks API.",
    "version": "2.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/api/v2",
  "paths": {
...
    "/tasks/{taskId}/time-entries": {
      "get": {
        "description": "Returns all time entries logged against the specified task, together with\nan aggregated total and the task's estimated duration.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "List time entries for a task",
        "operationId": "listTaskTimeEntries",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Time entries with aggregate summary.",
            "schema": {
              "$ref": "#/definitions/ResTaskTimeEntries"
            }
          },
          "400": {
            "description": "Invalid task ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "description": "Creates a manual time entry for the specified task without using the start/stop timer flow.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Log manual time entry",
        "operationId": "logManualTime",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReqLogManualTime"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Manual time entry created.",
            "schema": {
              "$ref": "#/definitions/ResTimeEntry"
            }
          },
          "400": {
            "description": "Validation error or invalid task ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/tasks/{taskId}/timer/discard": {
      "post": {
        "description": "Cancels and discards the running timer without creating a time entry.\nAny elapsed time is permanently lost.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Discard the active timer",
        "operationId": "discardTimer",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Timer discarded."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/tasks/{taskId}/timer/start": {
      "post": {
        "description": "Begins a new timer session for the specified task. Only one timer may be active per user at a time.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Start a task timer",
        "operationId": "startTimer",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the task to time.",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Timer started.",
            "schema": {
              "$ref": "#/definitions/ResActiveTimer"
            }
          },
          "400": {
            "description": "Invalid task ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error (e.g. timer already running)."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/tasks/{taskId}/timer/stop": {
      "post": {
        "description": "Stops the currently running timer for the task and converts it to a time entry.\nAn optional description and/or duration override may be provided.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Stop a task timer",
        "operationId": "stopTimer",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "taskId",
            "in": "path",
            "required": true
          },
          {
            "description": "Optional stop parameters. Body may be omitted entirely.",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ReqStopTimer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timer stopped and time entry created.",
            "schema": {
              "$ref": "#/definitions/ResTimeEntry"
            }
          },
          "400": {
            "description": "Invalid task ID or malformed body."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/time-entries/{entryId}": {
      "delete": {
        "description": "Permanently removes the specified time entry. Only the user who created the entry may delete it.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Delete a time entry",
        "operationId": "deleteTimeEntry",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the time entry.",
            "name": "entryId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Time entry deleted."
          },
          "400": {
            "description": "Invalid entry ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/timer/active": {
      "get": {
        "description": "Returns the currently running timer for the authenticated user, or null if no timer is active.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Get the active timer",
        "operationId": "getActiveTimer",
        "responses": {
          "200": {
            "description": "The active timer, or null.",
            "schema": {
              "$ref": "#/definitions/ResActiveTimer"
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/workspaces/{workspaceId}/projects/{projectId}/time-analytics": {
      "get": {
        "description": "Returns aggregated time tracking statistics for all tasks within the specified project.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "Get project time analytics",
        "operationId": "getProjectAnalytics",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "projectId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Project time analytics data."
          },
          "400": {
            "description": "Invalid project ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/projects/{projectId}/time-entries": {
      "get": {
        "description": "Returns all time entries logged against tasks in the project. Supports filtering by\nuser, start date, and end date.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "timeTracking"
        ],
        "summary": "List project time entries",
        "operationId": "listProjectTimeEntries",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "description": "Filter entries by a specific user UUID.",
            "name": "user_id",
            "in": "query"
          },
          {
            "type": "string",
            "format": "date-time",
            "description": "Filter entries starting from this RFC3339 timestamp (inclusive).",
            "name": "start_date",
            "in": "query"
          },
          {
            "type": "string",
            "format": "date-time",
            "description": "Filter entries up to this RFC3339 timestamp (inclusive).",
            "name": "end_date",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of time entries with task context.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResTimeEntryWithTask"
              }
            }
          },
          "400": {
            "description": "Invalid project ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ReqLogManualTime": {
      "type": "object",
      "title": "ReqLogManualTime defines the request payload for manually logging time on a task.",
      "required": [
        "duration_minutes"
      ],
      "properties": {
        "date": {
          "description": "Date the work was performed (YYYY-MM-DD). Defaults to today if omitted.",
          "type": "string",
          "x-go-name": "Date",
          "example": "2025-11-01"
        },
        "description": {
          "description": "Optional description of work done.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Code review and feedback"
        },
        "duration_minutes": {
          "description": "Duration of work in minutes.",
          "type": "integer",
          "format": "int64",
          "minimum": 1,
          "x-go-name": "DurationMinutes",
          "example": 60
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ResTaskTimeEntries": {
      "type": "object",
      "title": "ResTaskTimeEntries defines the aggregated response for time entries on a task.",
      "properties": {
        "entries": {
          "description": "List of individual time entries.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResTimeEntry"
          },
          "x-go-name": "Entries"
        },
        "estimated_minutes": {
          "description": "Task's estimated effort in minutes. Null if not set.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "EstimatedMinutes",
          "example": 300
        },
        "total_logged_minutes": {
          "description": "Total logged time in minutes across all entries.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "TotalLoggedMinutes",
          "example": 240
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ResTimeEntry": {
      "type": "object",
      "title": "ResTimeEntry defines the response payload for a time entry.",
      "properties": {
        "created_at": {
          "description": "Timestamp when the entry record was created.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt",
          "example": "2025-11-01T10:30:05Z"
        },
        "description": {
          "description": "Description of work performed.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Reviewed PR and added test cases"
        },
        "duration_minutes": {
          "description": "Total duration in minutes. Null if the timer is still running.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "DurationMinutes",
          "example": 90
        },
        "end_time": {
          "description": "Timer end time. Null if the timer is still running.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "EndTime",
          "example": "2025-11-01T10:30:00Z"
        },
        "id": {
          "description": "Unique time entry identifier.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "is_manual": {
          "description": "Whether this entry was manually logged (vs. started/stopped via timer).",
          "type": "boolean",
          "x-go-name": "IsManual",
          "example": false
        },
        "start_time": {
          "description": "Timer start time (RFC3339).",
          "type": "string",
          "format": "date-time",
          "x-go-name": "StartTime",
          "example": "2025-11-01T09:00:00Z"
        },
        "task_id": {
          "description": "UUID of the task this entry is associated with.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "TaskID",
          "example": "550e8400-e29b-41d4-a716-446655440001"
        },
        "user_full_name": {
          "description": "Full name of the user who logged the time.",
          "type": "string",
          "x-go-name": "UserFullName",
          "example": "Jane Doe"
        },
        "user_id": {
          "description": "UUID of the user who logged the time.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID",
          "example": "550e8400-e29b-41d4-a716-446655440002"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
  },
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "description": "A long-lived API key generated via the `/api/v2/auth/api-keys` endpoint. Intended for machine-to-machine access.",
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    },
    "cookieAuth": {
      "description": "HTTP-only session cookie set by the login/register endpoints. The cookie name is `nexus_user` and contains a signed JWT.\n",
      "type": "apiKey",
      "name": "Cookie",
      "in": "header"
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ]
}